Add filter cwcfp_change_disabled_to_readonly
Added filter cwcfp_change_disabled_to_readonly to allow the option for
changing text and textarea inputs to readonly instead of disabled.
This keeps the plugin working as it always has for the majority of
users, which prevents unexpected output. It also provides the option for
other users to switch to readonly and allow field values to be submitted
on checkout.